Clase 6/Laboratorio 1/requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
